// frontend-main/frontend-main/src/types/auth_type.ts
import type { BaseResponse } from ".";

////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
// LOGIN
////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
export interface LoginData {
  email: string;
  password: string;
}

export interface LoginResponse extends BaseResponse<{
    access: string;
    refresh: string;
    user: {
        email: string;
        full_name: string;
    }
}> {}